Move to vitalsValidation
